select data/count limit 100 from mysql,h2.oracle

+++ chuxue-commons-utils/src/main/java/org/chuxue/application/common/base/Pagination.java
@@ -14,7 +14,7 @@ import java.util.Map;
  * 版 本 ： V1.0
  */
 public class Pagination<T> implements Serializable {
-
+	
 	/**
 	 * @Fields serialVersionUID : TODO(用一句话描述这个变量表示什么)
 	 */
@@ -23,28 +23,28 @@ public class Pagination<T> implements Serializable {
 	protected Integer					pageNumber			= 1;
 	// 每页数据量大小
 	protected Integer					pageSize			= 10;
-
-	protected Integer					totalElements;
-
+	
+	protected Long						totalElements;
+	
 	// 查询数据id
 	protected String					uuid;
 	// 用户
 	protected String					username;
-
+	
 	protected List<T>					list;
 	protected List<SearchParameters>	searchList;
 	protected List<SortParameters>		sortList;
 	protected T							info;
 	protected Map<String, String>		map					= new HashMap<>();
-
+	
 	public Pagination() {
 	}
-
-	public Pagination(List<T> page, Integer total) {
+	
+	public Pagination(List<T> page, Long total) {
 		this.list = page;
 		this.totalElements = total;
 	}
-
+	
 	/**
 	 * 方法名 ： getInfo
 	 * 功 能 ： 返回变量 info 的值
@@ -54,7 +54,7 @@ public class Pagination<T> implements Serializable {
 	public T getInfo() {
 		return info;
 	}
-
+	
 	/**
 	 * 方法名 ： setInfo
 	 * 功 能 ： 设置变量 info 的值
@@ -62,7 +62,7 @@ public class Pagination<T> implements Serializable {
 	public void setInfo(T info) {
 		this.info = info;
 	}
-
+	
 	/**
 	 * 方法名 ： getPageNumber
 	 * 功 能 ： 返回变量 pageNumber 的值
@@ -72,7 +72,7 @@ public class Pagination<T> implements Serializable {
 	public Integer getPageNumber() {
 		return pageNumber;
 	}
-
+	
 	/**
 	 * 方法名 ： setPageNumber
 	 * 功 能 ： 设置变量 pageNumber 的值
@@ -80,7 +80,7 @@ public class Pagination<T> implements Serializable {
 	public void setPageNumber(Integer pageNumber) {
 		this.pageNumber = pageNumber;
 	}
-
+	
 	/**
 	 * 方法名 ： getPageSize
 	 * 功 能 ： 返回变量 pageSize 的值
@@ -90,7 +90,7 @@ public class Pagination<T> implements Serializable {
 	public Integer getPageSize() {
 		return pageSize;
 	}
-
+	
 	/**
 	 * 方法名 ： setPageSize
 	 * 功 能 ： 设置变量 pageSize 的值
@@ -98,7 +98,7 @@ public class Pagination<T> implements Serializable {
 	public void setPageSize(Integer pageSize) {
 		this.pageSize = pageSize;
 	}
-
+	
 	/**
 	 * 方法名 ： getUuid
 	 * 功 能 ： 返回变量 uuid 的值
@@ -108,7 +108,7 @@ public class Pagination<T> implements Serializable {
 	public String getUuid() {
 		return uuid;
 	}
-
+	
 	/**
 	 * 方法名 ： setUuid
 	 * 功 能 ： 设置变量 uuid 的值
@@ -116,7 +116,7 @@ public class Pagination<T> implements Serializable {
 	public void setUuid(String uuid) {
 		this.uuid = uuid;
 	}
-
+	
 	/**
 	 * 方法名 ： getList
 	 * 功 能 ： 返回变量 list 的值
@@ -126,7 +126,7 @@ public class Pagination<T> implements Serializable {
 	public List<T> getList() {
 		return list;
 	}
-
+	
 	/**
 	 * 方法名 ： setList
 	 * 功 能 ： 设置变量 list 的值
@@ -134,7 +134,7 @@ public class Pagination<T> implements Serializable {
 	public void setList(List<T> list) {
 		this.list = list;
 	}
-
+	
 	/**
 	 * 方法名 ： getUsername
 	 * 功 能 ： 返回变量 username 的值
@@ -144,7 +144,7 @@ public class Pagination<T> implements Serializable {
 	public String getUsername() {
 		return username;
 	}
-
+	
 	/**
 	 * 方法名 ： setUsername
 	 * 功 能 ： 设置变量 username 的值
@@ -152,11 +152,11 @@ public class Pagination<T> implements Serializable {
 	public void setUsername(String username) {
 		this.username = username;
 	}
-
+	
 	public Map<String, String> getMap() {
 		return map;
 	}
-
+	
 	/**
 	 * 方法名 ： setMap
 	 * 功 能 ： 设置变量 map 的值
@@ -164,7 +164,7 @@ public class Pagination<T> implements Serializable {
 	public void setMap(Map<String, String> map) {
 		this.map = map;
 	}
-
+	
 	/**
 	 * 方法名 ： getSearchList
 	 * 功 能 ： 返回变量 searchList 的值
@@ -174,7 +174,7 @@ public class Pagination<T> implements Serializable {
 	public List<SearchParameters> getSearchList() {
 		return searchList;
 	}
-
+	
 	/**
 	 * 方法名 ： setSearchList
 	 * 功 能 ： 设置变量 searchList 的值
@@ -182,7 +182,7 @@ public class Pagination<T> implements Serializable {
 	public void setSearchList(List<SearchParameters> searchList) {
 		this.searchList = searchList;
 	}
-
+	
 	/**
 	 * 方法名 ： getSortList
 	 * 功 能 ： 返回变量 sortList 的值
@@ -192,7 +192,7 @@ public class Pagination<T> implements Serializable {
 	public List<SortParameters> getSortList() {
 		return sortList;
 	}
-
+	
 	/**
 	 * 方法名 ： setSortList
 	 * 功 能 ： 设置变量 sortList 的值
@@ -200,25 +200,25 @@ public class Pagination<T> implements Serializable {
 	public void setSortList(List<SortParameters> sortList) {
 		this.sortList = sortList;
 	}
-	
+
 	/**
 	 * 方法名 ： getTotalElements
 	 * 功 能 ： 返回变量 totalElements 的值
 	 *
 	 * @return: Integer
 	 */
-	public Integer getTotalElements() {
+	public Long getTotalElements() {
 		return totalElements;
 	}
-	
+
 	/**
 	 * 方法名 ： setTotalElements
 	 * 功 能 ： 设置变量 totalElements 的值
 	 */
-	public void setTotalElements(Integer totalElements) {
+	public void setTotalElements(Long totalElements) {
 		this.totalElements = totalElements;
 	}
-	
+
 	/**
 	 * 方法名 ： toString
 	 * 功 能 ： TODO(这里用一句话描述这个方法的作用)
@@ -226,10 +226,10 @@ public class Pagination<T> implements Serializable {
 	 * 参 考 ： @see java.lang.Object#toString()
 	 * 作 者 ： Administrator
 	 */
-
+	
 	@Override
 	public String toString() {
 		return "Pagination [pageNumber=" + pageNumber + ", pageSize=" + pageSize + ", uuid=" + uuid + ", username=" + username + ", list=" + list + ", info=" + info + ", map=" + map + "]";
 	}
-
+	
 }
